requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
openai==1.0.0
pydantic==2.11.7
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
